Deprecated our first function.
Added CoreAPIV2-derecated.r
renamed ODATAcleanName.r to odataCleanName.r
copied and deprecated the function in favor of the correctly named function.
Moved the old tests with underscores to a separate file until we are ready to package and release. They should be deleted by then. There is a separate ticket for this.