RSDK-96 enabled testing against 5.3.9 and 6.0.2 PFS envs
Deleted duplicate auth and data files
Added symlinks to Auth/Data-2.7.1.json

To save on developers repeating update operations for the Auth and
Data JSON files used in automated testing of the package, have
updated the repo to instead use symlinks of the Auth-2.7.1.json
and Data-2.7.1.json files. In this way, a developer only need update
the Data-2.7.1.json file for the change to apply to other versions
of PFS being tested against.

The procedure used to create the symlink is
$ ln -s tests/testthat/test_environment/Auth-2.7.1.json \
    tests/testthat/test_environment/Auth-2.8.1.json
$ git add -f tests/testthat/test_environment/Auth-2.8.1.json

...and so on for the other Auth and Data files.